slight bug fix to damage
Damage listener was called before damaged was apply so it didnt feel correct. Also hazards were on a cooldown timer and so player was able to run through without getting damaged. Now every time the first object jumps in the hazard it gets damaged and the hazard attack cooldown starts.